Platform: add support of HiSilicon Poplar platform
It adds initial support of HiSilicon Poplar platform.  With it, we can
flash Android images to eMMC with fastboot, and boot the images into
Android system.

It's a derivative of HiKey platform support with a lot of code reused
from there.  Most of functionality available on HiKey is supported by
Poplar, except that GPIO and SD slot are currently not supported yet.

The most outstanding difference to HiKey is that Poplar doesn't support
GPT partition but only MBR, and partitions are named in the same way
how U-Boot names MBR partitions, i.e. mmcsdaX.  That said, fastboot
flash command will work like below.

  $ fastboot flash mmcsda3 system.img

Signed-off-by: Zhufu Wengang <zhufuwengang@hisilicon.com>
Signed-off-by: Shawn Guo <shawn.guo@linaro.org>