Drivers/DwUsbDxe: use ep 2 as OUT endpoint
The DwUsbDxe driver is coded to use ep1 as both IN and OUT endpoint.
It works for platforms like Hikey, on which the USB device endpoint is
bidirectional.  But it doesn't work on platform like Poplar, where the
USB device endpoints have fixed direction, i.e. ep1 is IN and ep2 is
OUT.  Since USB device on Hikey has ep1 ~ ep15 all work as
bidirectional endpoints, let's change the driver use ep1 as IN and ep2
as OUT endpoint, so that it works for both Hikey and Poplar.

It requires a corresponding change in edk2 counterpart, the Out Endpoint
descriptor in
EmbeddedPkg/Drivers/AndroidFastbootTransportUsbDxe/FastbootTransportUsb.c.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>