Platforms/Poplar: fix arrow keys support
The key is in the receive FIFO in PL011 driver. This driver supports
of 1, 16 or 32 chars.

Arrow key is presented as 3 chars. For example, DOWN key is "~[B".
By default, PcdUartDefaultReceiveFifoDepth is set as 1 in
MdePkg/MdePkg.dec. It means that arrow key is truncated. So it can't
work.

After updating to 0, the FIFO depth will be set either 16 or 32 chars.
The logic is in below.

  HardwareFifoDepth = (PL011_UARTPID2_VER (MmioRead32 (UartBase + UARTPID2)) \
                       > PL011_VER_R1P4) \
                      ? 32 : 16 ;

This is a porting of HiKey commit 39049e00d3cb (Platforms/Hisilicon: fix
arrow keys in release build) for Poplar.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>